Implemented radix-sort and refine counting-sort.
Radix-sort is based on counting-sort, so I improved
counting-sort for more support. The API of couting-sort
is also changed to enable caller to specify a min and max
value. Radix-sort can also support negative number, this is
"amazing"! :)

I also expand the set of test data to include more numbers,
bigger numbers and negative numbers.