Optimize RBT's transforming process.
I think the transformation in my last commit that rotates the tree
when uncle is red is correct. However, apparently there's a more
convinient means that only changes colors without any rotatation,
which is included in this commit.

I also added the explanation of the transformation when parent is
a right child.

The next step is to write test code.