Built up a framework for the operations of all kinds of hashes.
An operation table manages the dispatching of the operations
(Direct Accessing, Closed Accessing and Open Accessing).

Perfect Hashing may be not a part of this operation and it may
have different APIs. I'll come back to it after the other types
are ready.

I used C99's new feature -- Desinated Initailizers in the operation
dispatching table. It's cool!