Merge hash types CA_MOD and CA_MULT.
Enabled the 3rd variable parameter of HashInit to accept
a pointer to a hashing functions.

I don't need to make the number of slots exactly the same
as the maximum number which can be hashed to. The last
few slots can be empty.

TODO: rehash if load factor goes beyond 0.75.